provide a fix for asyncio on Windows + Python pre-3.8 (#105)
* provide a fix for asyncio on Windows + Python pre-3.8
+ test
+ enable github actions test run on Python 3.9

* black

* exclude python 3.9 + windows from test runs as there is no such version supported on github actions

* reverse the workflow condition

* ...

* .....

* exclude instead of the condition ...

thanks @reqc ! :)

* limit the fix to python  3.6/3.7

* ....

* ah, exclue osx+py3.9 too...

* black

* fix
