+++ .github/workflows/tests.yml
@@ -15,11 +15,16 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        python-version: [3.6, 3.7, 3.8]
+        python-version: [3.6, 3.7, 3.8, 3.9]
         os:
           - ubuntu-latest
           - macos-latest
           - windows-latest
+        exclude:
+          - os: windows-latest
+            python-version: 3.9
+          - os: macos-latest
+            python-version: 3.9
       fail-fast: false
 
     steps:
