+++ examples/yacat/yacat.py
@@ -4,7 +4,7 @@ from datetime import timedelta
 import pathlib
 import sys
 
-from yapapi import Executor, Task, WorkContext
+from yapapi import Executor, Task, WorkContext, windows_event_loop_fix
 from yapapi.log import enable_default_logger, log_summary, log_event_repr  # noqa
 from yapapi.package import vm
 
@@ -142,7 +142,7 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     # This is only required when running on Windows with Python prior to 3.8:
-    utils.windows_event_loop_fix()
+    windows_event_loop_fix()
 
     enable_default_logger(log_file=args.log_file)
