+++ tests/test_yapapi.py
@@ -1,3 +1,4 @@
+import asyncio
 import toml
 from pathlib import Path
 
@@ -9,3 +10,14 @@ def test_version():
         pyproject = toml.loads(f.read())
 
     assert yapapi.__version__ == pyproject["tool"]["poetry"]["version"]
+
+
+def test_windows_event_loop_fix():
+    async def _asyncio_test():
+        await asyncio.create_subprocess_shell("")
+
+    yapapi.windows_event_loop_fix()
+
+    l = asyncio.get_event_loop()
+    t = l.create_task(_asyncio_test())
+    l.run_until_complete(t)
