+++ yapapi/__init__.py
@@ -1,4 +1,6 @@
 """Golem Python API."""
+import asyncio
+import sys
 import toml
 
 from pathlib import Path
@@ -19,6 +21,22 @@ def get_version() -> str:
         return pyproject["tool"]["poetry"]["version"]
 
     return get_distribution("yapapi").version
+
+
+def windows_event_loop_fix():
+    """
+    Set up asyncio to use ProactorEventLoop implementation for new event loops on Windows.
+
+    This work-around is only needed for Python 3.6 and 3.7.
+    With Python 3.8, `ProactorEventLoop` is already the default on Windows.
+    """
+
+    if sys.platform == "win32" and sys.version_info < (3, 8):
+
+        class _WindowsEventPolicy(asyncio.events.BaseDefaultEventLoopPolicy):
+            _loop_factory = asyncio.windows_events.ProactorEventLoop
+
+        asyncio.set_event_loop_policy(_WindowsEventPolicy())
 
 
 __version__: str = get_version()
